reorganize to fix unallowed import cycles between statediff and statediff/ipfs
