+++ statediff/builder/builder.go
@@ -17,7 +17,7 @@
 // Contains a batch of utility type declarations used by the tests. As the node
 // operates on unique types, a lot of them are needed to check various features.
 
-package statediff
+package builder
 
 import (
 	"github.com/ethereum/go-ethereum/common"
