+++ statediff/publisher/ipfs/adder.go
@@ -17,16 +17,42 @@
 // Contains a batch of utility type declarations used by the tests. As the node
 // operates on unique types, a lot of them are needed to check various features.
 
-package statediff_test
+package ipfs
 
 import (
-	"testing"
+	"context"
 
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
+	"github.com/ipfs/go-ipfs/core"
+	"github.com/ipfs/go-ipfs/repo/fsrepo"
+	ipld "gx/ipfs/QmcKKBwfz6FyQdHR2jsXrrF6XeSBXYL86anmWNewpFpoF5/go-ipld-format"
 )
 
-func TestStateDiff(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "StateDiff Suite")
+type Adder interface {
+	Add(node ipld.Node) error
+}
+
+type adder struct {
+	n   *core.IpfsNode
+	ctx context.Context
+}
+
+func (a adder) Add(node ipld.Node) error {
+	return a.n.DAG.Add(a.n.Context(), node) // For some reason DAG.Add method is not being exposed by the ipld.DAGService
+}
+
+func NewAdder(repoPath string) (*adder, error) {
+	r, err := fsrepo.Open(repoPath)
+	if err != nil {
+		return nil, err
+	}
+	ctx := context.Background()
+	cfg := &core.BuildCfg{
+		Online: false,
+		Repo:   r,
+	}
+	ipfsNode, err := core.NewNode(ctx, cfg)
+	if err != nil {
+		return nil, err
+	}
+	return &adder{n: ipfsNode, ctx: ctx}, nil
 }
